@@ -1,77 +1,46 @@
-import React, { useState } from 'react';
-import { Link } from 'react-router-dom';
-import { Navbar, Nav, Container, Modal, Tab } from 'react-bootstrap';
-import SignUpForm from './SignupForm';
-import LoginForm from './LoginForm';
+import { AppBar, Box, Button, Toolbar, Typography } from "@mui/material";
+import { AuthContext } from "../context/authContext";
+import { useContext } from "react";
+import { Link, useNavigate } from "react-router-dom";
 
-import Auth from '../utils/auth';
 
-const AppNavbar = () => {
-  // set modal display state
-  const [showModal, setShowModal] = useState(false);
+function Navbar () {
+  let navigate = useNavigate();
+  const { user, logout } = useContext(AuthContext);
+
+  const onLogout = () => {
+    logout();
+    navigate("/home");
+  };
 
   return (
-    <>
-      <Navbar bg='dark' variant='dark' expand='lg'>
-        <Container fluid>
-          <Navbar.Brand as={Link} to='/'>
-            Google Books Search
-          </Navbar.Brand>
-          <Navbar.Toggle aria-controls='navbar' />
-          <Navbar.Collapse id='navbar' className='d-flex flex-row-reverse'>
-            <Nav className='ml-auto d-flex'>
-              <Nav.Link as={Link} to='/'>
-                Search For Books
-              </Nav.Link>
-              {/* if user is logged in show saved books and logout */}
-              {Auth.loggedIn() ? (
-                <>
-                  <Nav.Link as={Link} to='/saved'>
-                    See Your Books
-                  </Nav.Link>
-                  <Nav.Link onClick={Auth.logout}>Logout</Nav.Link>
-                </>
-              ) : (
-                <Nav.Link onClick={() => setShowModal(true)}>Login/Sign Up</Nav.Link>
-              )}
-            </Nav>
-          </Navbar.Collapse>
-        </Container>
-      </Navbar>
-      {/* set modal data up */}
-      <Modal
-        size='lg'
-        show={showModal}
-        onHide={() => setShowModal(false)}
-        aria-labelledby='signup-modal'>
-        {/* tab container to do either signup or login component */}
-        <Tab.Container defaultActiveKey='login'>
-          <Modal.Header closeButton>
-            <Modal.Title id='signup-modal'>
-              <Nav variant='pills'>
-                <Nav.Item>
-                  <Nav.Link eventKey='login'>Login</Nav.Link>
-                </Nav.Item>
-                <Nav.Item>
-                  <Nav.Link eventKey='signup'>Sign Up</Nav.Link>
-                </Nav.Item>
-              </Nav>
-            </Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            <Tab.Content>
-              <Tab.Pane eventKey='login'>
-                <LoginForm handleModalClose={() => setShowModal(false)} />
-              </Tab.Pane>
-              <Tab.Pane eventKey='signup'>
-                <SignUpForm handleModalClose={() => setShowModal(false)} />
-              </Tab.Pane>
-            </Tab.Content>
-          </Modal.Body>
-        </Tab.Container>
-      </Modal>
-    </>
+    <Box sx={{ flexGrow: 1 }}>
+      <AppBar position="static">
+        <Toolbar>
+          <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+            <Link to="/home" style={{ textDecoration: "none", color: "white" }}>Home </Link>
+            <Link to="/login" style={{ textDecoration: "none", color: "white" }}> Login</Link>
+            <Link to="/signup" style={{ textDecoration: "none", color: "white" }}> Register</Link>
+            <Link to="/searchBooks" style={{ textDecoration: "none", color: "white" }}> SearchBooks</Link>
+            <Link to="/savedBooks" style={{ textDecoration: "none", color: "white" }}> SavedBooks</Link>
+          </Typography>
+          <Box alightitems="right" sx={{ flexGrow: 1, textAlign: "right" }}>
+            {/** If user is logged in, show logout button, else show login and register buttons */}
+            {user ? 
+              <>
+                <Button style={{ textDecoration: "none", color: "white" }} onClick={onLogout}>Logout</Button>
+              </>
+              :
+              <>
+                <Button style={{ textDecoration: "none", color: "white" }} component={Link} to="/login">Login</Button>
+                <Button style={{ textDecoration: "none", color: "white" }} component={Link} to="/signup">Register</Button>
+              </>
+            }
+          </Box>
+        </Toolbar>
+      </AppBar>
+    </Box>
   );
 };
 
-export default AppNavbar;
+export default Navbar;
