@@ -1,154 +1,75 @@
-import React, { useState, useEffect } from 'react';
-import {
-  Container,
-  Col,
-  Form,
-  Button,
-  Card,
-  Row
-} from 'react-bootstrap';
+import React from 'react';
+import SaveBook from '../components/SaveBookBtn'
+import { Alert, Container, Box, TextField, Button, Typography, Link, List, ListItem } from '@mui/material';
+import { useForm } from '../utils/hooks';
+import { useLazyQuery } from '@apollo/client';
+import { SEARCH_BOOKS } from '../utils/queries';
 
-import Auth from '../utils/auth';
-import { saveBook, searchGoogleBooks } from '../utils/API';
-import { saveBookIds, getSavedBookIds } from '../utils/localStorage';
+function SearchBooks() {
+  const [searchBooks, { loading, error, data }] = useLazyQuery(SEARCH_BOOKS);
 
-const SearchBooks = () => {
-  // create state for holding returned google api data
-  const [searchedBooks, setSearchedBooks] = useState([]);
-  // create state for holding our search field data
-  const [searchInput, setSearchInput] = useState('');
+  const search = () => {
+    searchBooks({ variables: { query: values.searchInput } });
+  };
 
-  // create state to hold saved bookId values
-  const [savedBookIds, setSavedBookIds] = useState(getSavedBookIds());
-
-  // set up useEffect hook to save `savedBookIds` list to localStorage on component unmount
-  // learn more here: https://reactjs.org/docs/hooks-effect.html#effects-with-cleanup
-  useEffect(() => {
-    return () => saveBookIds(savedBookIds);
+  const { onChange, onSubmit, values } = useForm(search, {
+    searchInput: '',
   });
 
-  // create method to search for books and set state on form submit
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
-
-    if (!searchInput) {
-      return false;
-    }
-
-    try {
-      const response = await searchGoogleBooks(searchInput);
-
-      if (!response.ok) {
-        throw new Error('something went wrong!');
-      }
-
-      const { items } = await response.json();
-
-      const bookData = items.map((book) => ({
-        bookId: book.id,
-        authors: book.volumeInfo.authors || ['No author to display'],
-        title: book.volumeInfo.title,
-        description: book.volumeInfo.description,
-        image: book.volumeInfo.imageLinks?.thumbnail || '',
-      }));
-
-      setSearchedBooks(bookData);
-      setSearchInput('');
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  // create function to handle saving a book to our database
-  const handleSaveBook = async (bookId) => {
-    // find the book in `searchedBooks` state by the matching id
-    const bookToSave = searchedBooks.find((book) => book.bookId === bookId);
-
-    // get token
-    const token = Auth.loggedIn() ? Auth.getToken() : null;
-
-    if (!token) {
-      return false;
-    }
-
-    try {
-      const response = await saveBook(bookToSave, token);
-
-      if (!response.ok) {
-        throw new Error('something went wrong!');
-      }
-
-      // if book successfully saves to user's account, save book id to state
-      setSavedBookIds([...savedBookIds, bookToSave.bookId]);
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
   return (
-    <>
-      <div className='text-light bg-dark pt-5'>
-        <Container>
-          <h1>Search for Books!</h1>
-          <Form onSubmit={handleFormSubmit}>
-            <Row>
-              <Col xs={12} md={8}>
-                <Form.Control
-                  name='searchInput'
-                  value={searchInput}
-                  onChange={(e) => setSearchInput(e.target.value)}
-                  type='text'
-                  size='lg'
-                  placeholder='Search for a book'
-                />
-              </Col>
-              <Col xs={12} md={4}>
-                <Button type='submit' variant='success' size='lg'>
-                  Submit Search
-                </Button>
-              </Col>
-            </Row>
-          </Form>
-        </Container>
-      </div>
+    <Container maxWidth="md">
+      <Box my={4}>
+        <Typography variant="h4" component="h1" gutterBottom>
+          Search for Books!
+        </Typography>
+        <form onSubmit={onSubmit}>
+          <TextField
+            fullWidth
+            name='searchInput'
+            label="Search for a book"
+            value={values.searchInput}
+            onChange={onChange}
+          />
+          <Button
+            variant="contained"
+            color="primary"
+            type="submit"
+            onClick={onSubmit}
+          >
+            Submit Book Search
+          </Button>
+        </form>
+      </Box>
+      <Box my={4}>
 
-      <Container>
-        <h2 className='pt-5'>
-          {searchedBooks.length
-            ? `Viewing ${searchedBooks.length} results:`
-            : 'Search for a book to begin'}
-        </h2>
-        <Row>
-          {searchedBooks.map((book) => {
-            return (
-              <Col md="4">
-                <Card key={book.bookId} border='dark'>
-                  {book.image ? (
-                    <Card.Img src={book.image} alt={`The cover for ${book.title}`} variant='top' />
-                  ) : null}
-                  <Card.Body>
-                    <Card.Title>{book.title}</Card.Title>
-                    <p className='small'>Authors: {book.authors}</p>
-                    <Card.Text>{book.description}</Card.Text>
-                    {Auth.loggedIn() && (
-                      <Button
-                        disabled={savedBookIds?.some((savedBookId) => savedBookId === book.bookId)}
-                        className='btn-block btn-info'
-                        onClick={() => handleSaveBook(book.bookId)}>
-                        {savedBookIds?.some((savedBookId) => savedBookId === book.bookId)
-                          ? 'This book has already been saved!'
-                          : 'Save this Book!'}
-                      </Button>
-                    )}
-                  </Card.Body>
-                </Card>
-              </Col>
-            );
-          })}
-        </Row>
-      </Container>
-    </>
+        {loading && <Typography variant="body1">Loading...</Typography>}
+
+        {error && <Alert severity="error" sx={{ mt: 2 }}>
+          An error occurred: {error.message}
+        </Alert>}
+
+        {data?.searchBooks && (
+          <List>
+            {data.searchBooks.map((book) => (
+              <ListItem key={book.bookId}>
+                <Link href={book.link}>
+                  <img src={book.image} alt={book.title} />
+                  <Typography variant="body1">
+                    <strong>Description:</strong> {book.description}
+                  </Typography>
+                  <Typography variant="body1">
+                    <strong>Authors:</strong> {book.authors.join(', ')}
+                  </Typography>
+                </Link>
+                <SaveBook book={book} />
+              </ListItem>
+            ))}
+          </List>
+
+        )}
+
+      </Box>
+    </Container>
   );
-};
-
+}
 export default SearchBooks;
