@@ -1,4 +1,4 @@
-const { Schema, model } = require('mongoose');
+const { model, Schema } = require('mongoose');
 
 // This is a subdocument schema, it won't become its own model but we'll use it as the schema for the User's `savedBooks` array in User.js
 const bookSchema = new Schema({
@@ -9,7 +9,7 @@ const bookSchema = new Schema({
   ],
   description: {
     type: String,
-    required: true,
+    required: false,
   },
   // saved book id from GoogleBooks
   bookId: {
@@ -28,6 +28,5 @@ const bookSchema = new Schema({
   },
 });
 
-const Book = model('Book', bookSchema);
+module.exports = model('Book', bookSchema);
 
-module.exports = Book;
